Add refines command
This is a shell wrapper around FDR's refines command [1].  We run it
inside of a Docker container (created via `make`) so that you don't have
to worry about how to install FDR on your host OS.  We expose $PWD
inside the container so that you can pass in whichever CSP scripts you
want to analyze.  We also expose FDR's per-user config directory, so
that we can save your FDR license file on the host, allowing you to
reuse it however many times you spin up an FDR container.

[1] http://www.cs.ox.ac.uk/projects/fdr/manual/command_line.html